+++ packages/interactive_media_ads/pigeons/interactive_media_ads_android.dart
@@ -6,7 +6,7 @@
 // https://github.com/flutter/packages/pull/6371 lands. This file uses the
 // Kotlin ProxyApi feature from pigeon.
 // ignore_for_file: avoid_unused_constructor_parameters
-/*
+
 import 'package:pigeon/pigeon.dart';
 
 @ConfigurePigeon(
@@ -327,7 +327,15 @@ abstract class AdsRequest {
         'com.google.ads.interactivemedia.v3.api.player.ContentProgressProvider',
   ),
 )
-abstract class ContentProgressProvider {}
+abstract class ContentProgressProvider {
+  ContentProgressProvider();
+
+  /// Sets an update on the progress of the video.
+  ///
+  /// This is a custom method added to the native class because the native
+  /// method `getContentProgress` requires a synchronous return value.
+  void setContentProgress(VideoProgressUpdate update);
+}
 
 /// An object which handles playing ads after they've been received from the
 /// server.
@@ -738,4 +746,3 @@ abstract class AdEventListener {
   /// Respond to an occurrence of an AdEvent.
   late final void Function(AdEvent event) onAdEvent;
 }
-*/