+++ packages/interactive_media_ads/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.2.1
+
+* Adds internal wrapper for Android native `ContentProgressProvider`.
+
+## 0.2.0
+
+* Adds support for pausing and resuming Ad playback. See `AdsManager.pause` and `AdsManager.resume`.
+* Adds support to skip an Ad. See `AdsManager.skip` and `AdsManager.discardAdBreak`.
+* **Breaking Change** To keep platform consistency, Android no longer continues playing an Ad 
+  whenever it returns from an Ad click. Call `AdsManager.resume` to resume Ad playback.
+
 ## 0.1.2+6
 
 * Fixes bug where the ad would play when the app returned to foreground during content playback.